Add logging.py and displace AuditLogger there
