+++ shared/hsm.py
@@ -16,7 +16,7 @@ from ubinascii import hexlify as b2a_hex
 from ubinascii import unhexlify as a2b_hex
 from uhashlib import sha256
 from ucollections import OrderedDict
-from files import CardSlot, CardMissingError
+from logging import AuditLogger
 from serializations import CTxOut
 
 # where we save policy/config
@@ -401,54 +401,6 @@ class ApprovalRule:
                 assert o.amount == wanted, 'not all output amounts are equal'
 
         return True
-
-class AuditLogger:
-    def __init__(self, dirname, digest, never_log):
-        self.dirname = dirname
-        self.digest = digest
-        self.never_log = never_log
-
-    def __enter__(self):
-        try:
-            if self.never_log:
-                raise NotImplementedError
-
-            self.card = CardSlot().__enter__()
-
-            d  = self.card.get_sd_root() + '/' + self.dirname
-
-            # mkdir if needed
-            try: uos.stat(d)
-            except: uos.mkdir(d)
-                
-            self.fname = d + '/' + b2a_hex(self.digest[-8:]).decode('ascii') + '.log'
-            self.fd = open(self.fname, 'a+t')       # append mode
-        except (CardMissingError, OSError, NotImplementedError):
-            # may be fatal or not, depending on configuration
-            self.fname = self.card = None
-            self.fd = sys.stdout
-
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        if exc_value:
-            self.fd.write('\n\n---- Coldcard Exception ----\n')
-            sys.print_exception(exc_value, self.fd)
-
-        self.fd.write('\n===\n\n')
-
-        if self.card:
-            assert self.fd != sys.stdout
-            self.fd.close()
-            self.card.__exit__(exc_type, exc_value, traceback)
-
-    @property
-    def is_unsaved(self):
-        return not self.card
-
-    def info(self, msg):
-        print(msg, file=self.fd)
-        #if self.fd != sys.stdout: print(msg)
 
 class HSMPolicy:
     # implements and enforces the HSM signing/activity/logging policy
